Move poolsense base entity to its own file
