+++ .coveragerc
@@ -979,6 +979,7 @@ omit =
     homeassistant/components/poolsense/__init__.py
     homeassistant/components/poolsense/binary_sensor.py
     homeassistant/components/poolsense/coordinator.py
+    homeassistant/components/poolsense/entity.py
     homeassistant/components/poolsense/sensor.py
     homeassistant/components/powerwall/__init__.py
     homeassistant/components/progettihwsw/__init__.py
